Split activity.py and engine.py into packages
activity.py (3,334 lines) -> activity/ package:
  - generator.py: ActivityGenerator class + data constants (2,837 lines)
  - network.py: Network helper functions and IP/hostname data (397 lines)
  - helpers.py: RNG, OS detection, command parameterization (131 lines)
  - __init__.py: Re-exports for backward-compatible imports

engine.py (2,711 lines) -> engine/ package with mixin architecture:
  - core.py: GenerationEngine class (inherits 3 mixins), init, generate()
  - baseline.py: BaselineMixin (hour-by-hour loop, system traffic)
  - storyline.py: StorylineMixin (event execution, supplementary events)
  - emitter_setup.py: EmitterSetupMixin (emitter init, SID registry, AD)
  - __init__.py: Re-exports GenerationEngine

962 tests pass (1 pre-existing skip). No import changes needed in
consuming code — package __init__.py files maintain the same public API.
