+++ src/evidenceforge/generation/engine/__init__.py
@@ -1,0 +1,5 @@
+"""Generation engine package for coordinated log production."""
+
+from .core import GenerationEngine
+
+__all__ = ["GenerationEngine"]
