+++ tests/unit/test_engine.py
@@ -24,11 +24,11 @@ class TestGenerationEngine:
         7 emitters. This fixture mocks the 5 new ones so existing tests
         that only patch WindowsEventEmitter and ZeekEmitter still work.
         """
-        with patch('evidenceforge.generation.engine.EcarEmitter') as m1, \
-             patch('evidenceforge.generation.engine.SyslogEmitter') as m2, \
-             patch('evidenceforge.generation.engine.BashHistoryEmitter') as m3, \
-             patch('evidenceforge.generation.engine.SnortEmitter') as m4, \
-             patch('evidenceforge.generation.engine.WebEmitter') as m5:
+        with patch('evidenceforge.generation.engine.emitter_setup.EcarEmitter') as m1, \
+             patch('evidenceforge.generation.engine.emitter_setup.SyslogEmitter') as m2, \
+             patch('evidenceforge.generation.engine.emitter_setup.BashHistoryEmitter') as m3, \
+             patch('evidenceforge.generation.engine.emitter_setup.SnortEmitter') as m4, \
+             patch('evidenceforge.generation.engine.emitter_setup.WebEmitter') as m5:
             yield m1, m2, m3, m4, m5
 
     @pytest.fixture
@@ -141,23 +141,23 @@ class TestGenerationEngine:
             ]
         )
 
-    @patch('evidenceforge.generation.engine.ActivityGenerator')
-    @patch('evidenceforge.generation.engine.ZeekReporterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPacketFilterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPeEmitter')
-    @patch('evidenceforge.generation.engine.ZeekOcspEmitter')
-    @patch('evidenceforge.generation.engine.ZeekX509Emitter')
-    @patch('evidenceforge.generation.engine.ZeekWeirdEmitter')
-    @patch('evidenceforge.generation.engine.ZeekNtpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDhcpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekFilesEmitter')
-    @patch('evidenceforge.generation.engine.ZeekSslEmitter')
-    @patch('evidenceforge.generation.engine.ZeekHttpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDnsEmitter')
-    @patch('evidenceforge.generation.engine.ZeekEmitter')
-    @patch('evidenceforge.generation.engine.SysmonEventEmitter')
-    @patch('evidenceforge.generation.engine.WindowsEventEmitter')
-    @patch('evidenceforge.generation.engine.load_format')
+    @patch('evidenceforge.generation.engine.core.ActivityGenerator')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekReporterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPacketFilterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPeEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekOcspEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekX509Emitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekWeirdEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekNtpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDhcpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekFilesEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekSslEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekHttpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDnsEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.SysmonEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.WindowsEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.load_format')
     def test_initialize_creates_emitters(
         self, mock_load_format, mock_windows, mock_sysmon, mock_zeek, mock_zeek_dns,
         mock_zeek_http, mock_zeek_ssl, mock_zeek_files, mock_zeek_dhcp, mock_zeek_ntp, mock_zeek_weird, mock_zeek_x509, mock_zeek_ocsp, mock_zeek_pe, mock_zeek_pf, mock_zeek_reporter, mock_activity_gen,
@@ -182,23 +182,23 @@ class TestGenerationEngine:
         assert 'zeek_ssl' in engine.emitters
         assert 'zeek_files' in engine.emitters
 
-    @patch('evidenceforge.generation.engine.ActivityGenerator')
-    @patch('evidenceforge.generation.engine.ZeekReporterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPacketFilterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPeEmitter')
-    @patch('evidenceforge.generation.engine.ZeekOcspEmitter')
-    @patch('evidenceforge.generation.engine.ZeekX509Emitter')
-    @patch('evidenceforge.generation.engine.ZeekWeirdEmitter')
-    @patch('evidenceforge.generation.engine.ZeekNtpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDhcpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekFilesEmitter')
-    @patch('evidenceforge.generation.engine.ZeekSslEmitter')
-    @patch('evidenceforge.generation.engine.ZeekHttpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDnsEmitter')
-    @patch('evidenceforge.generation.engine.ZeekEmitter')
-    @patch('evidenceforge.generation.engine.WindowsEventEmitter')
-    @patch('evidenceforge.generation.engine.SysmonEventEmitter')
-    @patch('evidenceforge.generation.engine.load_format')
+    @patch('evidenceforge.generation.engine.core.ActivityGenerator')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekReporterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPacketFilterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPeEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekOcspEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekX509Emitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekWeirdEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekNtpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDhcpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekFilesEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekSslEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekHttpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDnsEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.WindowsEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.SysmonEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.load_format')
     def test_initialize_resolves_time_window(
         self, mock_load_format, mock_sysmon, mock_windows, mock_zeek, mock_zeek_dns, mock_zeek_http, mock_zeek_ssl, mock_zeek_files, mock_zeek_dhcp, mock_zeek_ntp, mock_zeek_weird, mock_zeek_x509, mock_zeek_ocsp, mock_zeek_pe, mock_zeek_pf, mock_zeek_reporter, mock_activity_gen,
         minimal_scenario, tmp_path
@@ -215,23 +215,23 @@ class TestGenerationEngine:
         assert engine.start_time == datetime(2024, 1, 15, 10, 0, 0, tzinfo=timezone.utc)
         assert engine.end_time == datetime(2024, 1, 15, 12, 0, 0, tzinfo=timezone.utc)
 
-    @patch('evidenceforge.generation.engine.ActivityGenerator')
-    @patch('evidenceforge.generation.engine.ZeekReporterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPacketFilterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPeEmitter')
-    @patch('evidenceforge.generation.engine.ZeekOcspEmitter')
-    @patch('evidenceforge.generation.engine.ZeekX509Emitter')
-    @patch('evidenceforge.generation.engine.ZeekWeirdEmitter')
-    @patch('evidenceforge.generation.engine.ZeekNtpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDhcpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekFilesEmitter')
-    @patch('evidenceforge.generation.engine.ZeekSslEmitter')
-    @patch('evidenceforge.generation.engine.ZeekHttpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDnsEmitter')
-    @patch('evidenceforge.generation.engine.ZeekEmitter')
-    @patch('evidenceforge.generation.engine.WindowsEventEmitter')
-    @patch('evidenceforge.generation.engine.SysmonEventEmitter')
-    @patch('evidenceforge.generation.engine.load_format')
+    @patch('evidenceforge.generation.engine.core.ActivityGenerator')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekReporterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPacketFilterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPeEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekOcspEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekX509Emitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekWeirdEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekNtpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDhcpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekFilesEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekSslEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekHttpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDnsEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.WindowsEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.SysmonEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.load_format')
     def test_initialize_creates_output_directory(
         self, mock_load_format, mock_sysmon, mock_windows, mock_zeek, mock_zeek_dns, mock_zeek_http, mock_zeek_ssl, mock_zeek_files, mock_zeek_dhcp, mock_zeek_ntp, mock_zeek_weird, mock_zeek_x509, mock_zeek_ocsp, mock_zeek_pe, mock_zeek_pf, mock_zeek_reporter, mock_activity_gen,
         minimal_scenario, tmp_path
@@ -247,23 +247,23 @@ class TestGenerationEngine:
 
         assert output_dir.exists()
 
-    @patch('evidenceforge.generation.engine.ActivityGenerator')
-    @patch('evidenceforge.generation.engine.ZeekReporterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPacketFilterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPeEmitter')
-    @patch('evidenceforge.generation.engine.ZeekOcspEmitter')
-    @patch('evidenceforge.generation.engine.ZeekX509Emitter')
-    @patch('evidenceforge.generation.engine.ZeekWeirdEmitter')
-    @patch('evidenceforge.generation.engine.ZeekNtpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDhcpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekFilesEmitter')
-    @patch('evidenceforge.generation.engine.ZeekSslEmitter')
-    @patch('evidenceforge.generation.engine.ZeekHttpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDnsEmitter')
-    @patch('evidenceforge.generation.engine.ZeekEmitter')
-    @patch('evidenceforge.generation.engine.WindowsEventEmitter')
-    @patch('evidenceforge.generation.engine.SysmonEventEmitter')
-    @patch('evidenceforge.generation.engine.load_format')
+    @patch('evidenceforge.generation.engine.core.ActivityGenerator')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekReporterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPacketFilterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPeEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekOcspEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekX509Emitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekWeirdEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekNtpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDhcpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekFilesEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekSslEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekHttpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDnsEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.WindowsEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.SysmonEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.load_format')
     def test_initialize_sets_state_manager_time(
         self, mock_load_format, mock_sysmon, mock_windows, mock_zeek, mock_zeek_dns, mock_zeek_http, mock_zeek_ssl, mock_zeek_files, mock_zeek_dhcp, mock_zeek_ntp, mock_zeek_weird, mock_zeek_x509, mock_zeek_ocsp, mock_zeek_pe, mock_zeek_pf, mock_zeek_reporter, mock_activity_gen,
         minimal_scenario, tmp_path
@@ -384,23 +384,23 @@ class TestGenerationEngine:
 
         assert times == []
 
-    @patch('evidenceforge.generation.engine.ActivityGenerator')
-    @patch('evidenceforge.generation.engine.ZeekReporterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPacketFilterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPeEmitter')
-    @patch('evidenceforge.generation.engine.ZeekOcspEmitter')
-    @patch('evidenceforge.generation.engine.ZeekX509Emitter')
-    @patch('evidenceforge.generation.engine.ZeekWeirdEmitter')
-    @patch('evidenceforge.generation.engine.ZeekNtpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDhcpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekFilesEmitter')
-    @patch('evidenceforge.generation.engine.ZeekSslEmitter')
-    @patch('evidenceforge.generation.engine.ZeekHttpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDnsEmitter')
-    @patch('evidenceforge.generation.engine.ZeekEmitter')
-    @patch('evidenceforge.generation.engine.WindowsEventEmitter')
-    @patch('evidenceforge.generation.engine.SysmonEventEmitter')
-    @patch('evidenceforge.generation.engine.load_format')
+    @patch('evidenceforge.generation.engine.core.ActivityGenerator')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekReporterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPacketFilterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPeEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekOcspEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekX509Emitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekWeirdEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekNtpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDhcpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekFilesEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekSslEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekHttpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDnsEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.WindowsEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.SysmonEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.load_format')
     def test_generate_baseline_filters_enabled_users(
         self, mock_load_format, mock_sysmon, mock_windows, mock_zeek, mock_zeek_dns, mock_zeek_http, mock_zeek_ssl, mock_zeek_files, mock_zeek_dhcp, mock_zeek_ntp, mock_zeek_weird, mock_zeek_x509, mock_zeek_ocsp, mock_zeek_pe, mock_zeek_pf, mock_zeek_reporter, mock_activity_gen,
         minimal_scenario, tmp_path
@@ -432,23 +432,23 @@ class TestGenerationEngine:
         # (or possibly zero times if no events generated due to randomness)
         assert mock_activity_instance.get_baseline_pattern.called
 
-    @patch('evidenceforge.generation.engine.ActivityGenerator')
-    @patch('evidenceforge.generation.engine.ZeekReporterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPacketFilterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPeEmitter')
-    @patch('evidenceforge.generation.engine.ZeekOcspEmitter')
-    @patch('evidenceforge.generation.engine.ZeekX509Emitter')
-    @patch('evidenceforge.generation.engine.ZeekWeirdEmitter')
-    @patch('evidenceforge.generation.engine.ZeekNtpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDhcpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekFilesEmitter')
-    @patch('evidenceforge.generation.engine.ZeekSslEmitter')
-    @patch('evidenceforge.generation.engine.ZeekHttpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDnsEmitter')
-    @patch('evidenceforge.generation.engine.ZeekEmitter')
-    @patch('evidenceforge.generation.engine.WindowsEventEmitter')
-    @patch('evidenceforge.generation.engine.SysmonEventEmitter')
-    @patch('evidenceforge.generation.engine.load_format')
+    @patch('evidenceforge.generation.engine.core.ActivityGenerator')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekReporterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPacketFilterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPeEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekOcspEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekX509Emitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekWeirdEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekNtpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDhcpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekFilesEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekSslEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekHttpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDnsEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.WindowsEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.SysmonEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.load_format')
     def test_generate_baseline_hour_by_hour(
         self, mock_load_format, mock_sysmon, mock_windows, mock_zeek, mock_zeek_dns, mock_zeek_http, mock_zeek_ssl, mock_zeek_files, mock_zeek_dhcp, mock_zeek_ntp, mock_zeek_weird, mock_zeek_x509, mock_zeek_ocsp, mock_zeek_pe, mock_zeek_pf, mock_zeek_reporter, mock_activity_gen,
         minimal_scenario, tmp_path
@@ -516,24 +516,24 @@ class TestGenerationEngine:
     # test_match_activity_to_events_* deleted in Phase 8.4
     # Keyword matching replaced by typed event declarations in scenario YAML
 
-    @patch('evidenceforge.generation.engine.GroundTruthGenerator')
-    @patch('evidenceforge.generation.engine.ActivityGenerator')
-    @patch('evidenceforge.generation.engine.ZeekReporterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPacketFilterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPeEmitter')
-    @patch('evidenceforge.generation.engine.ZeekOcspEmitter')
-    @patch('evidenceforge.generation.engine.ZeekX509Emitter')
-    @patch('evidenceforge.generation.engine.ZeekWeirdEmitter')
-    @patch('evidenceforge.generation.engine.ZeekNtpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDhcpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekFilesEmitter')
-    @patch('evidenceforge.generation.engine.ZeekSslEmitter')
-    @patch('evidenceforge.generation.engine.ZeekHttpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDnsEmitter')
-    @patch('evidenceforge.generation.engine.ZeekEmitter')
-    @patch('evidenceforge.generation.engine.WindowsEventEmitter')
-    @patch('evidenceforge.generation.engine.SysmonEventEmitter')
-    @patch('evidenceforge.generation.engine.load_format')
+    @patch('evidenceforge.generation.engine.core.GroundTruthGenerator')
+    @patch('evidenceforge.generation.engine.core.ActivityGenerator')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekReporterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPacketFilterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPeEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekOcspEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekX509Emitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekWeirdEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekNtpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDhcpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekFilesEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekSslEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekHttpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDnsEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.WindowsEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.SysmonEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.load_format')
     def test_execute_storyline_tracks_malicious_events(
         self, mock_load_format, mock_sysmon, mock_windows, mock_zeek, mock_zeek_dns, mock_zeek_http, mock_zeek_ssl, mock_zeek_files, mock_zeek_dhcp, mock_zeek_ntp, mock_zeek_weird, mock_zeek_x509, mock_zeek_ocsp, mock_zeek_pe, mock_zeek_pf, mock_zeek_reporter,
         mock_activity_gen, mock_gt_gen, scenario_with_storyline, tmp_path
@@ -558,24 +558,24 @@ class TestGenerationEngine:
         assert engine.malicious_events[0]['actor'] == 'attacker'
         assert engine.malicious_events[0]['system'] == 'TEST-01'
 
-    @patch('evidenceforge.generation.engine.GroundTruthGenerator')
-    @patch('evidenceforge.generation.engine.ActivityGenerator')
-    @patch('evidenceforge.generation.engine.ZeekReporterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPacketFilterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPeEmitter')
-    @patch('evidenceforge.generation.engine.ZeekOcspEmitter')
-    @patch('evidenceforge.generation.engine.ZeekX509Emitter')
-    @patch('evidenceforge.generation.engine.ZeekWeirdEmitter')
-    @patch('evidenceforge.generation.engine.ZeekNtpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDhcpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekFilesEmitter')
-    @patch('evidenceforge.generation.engine.ZeekSslEmitter')
-    @patch('evidenceforge.generation.engine.ZeekHttpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDnsEmitter')
-    @patch('evidenceforge.generation.engine.ZeekEmitter')
-    @patch('evidenceforge.generation.engine.WindowsEventEmitter')
-    @patch('evidenceforge.generation.engine.SysmonEventEmitter')
-    @patch('evidenceforge.generation.engine.load_format')
+    @patch('evidenceforge.generation.engine.core.GroundTruthGenerator')
+    @patch('evidenceforge.generation.engine.core.ActivityGenerator')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekReporterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPacketFilterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPeEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekOcspEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekX509Emitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekWeirdEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekNtpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDhcpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekFilesEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekSslEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekHttpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDnsEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.WindowsEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.SysmonEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.load_format')
     def test_generate_calls_ground_truth_when_malicious_events(
         self, mock_load_format, mock_sysmon, mock_windows, mock_zeek, mock_zeek_dns, mock_zeek_http, mock_zeek_ssl, mock_zeek_files, mock_zeek_dhcp, mock_zeek_ntp, mock_zeek_weird, mock_zeek_x509, mock_zeek_ocsp, mock_zeek_pe, mock_zeek_pf, mock_zeek_reporter,
         mock_activity_gen, mock_gt_gen, scenario_with_storyline, tmp_path
@@ -601,24 +601,24 @@ class TestGenerationEngine:
         assert mock_gt_gen.called
         assert mock_gt_instance.generate.called
 
-    @patch('evidenceforge.generation.engine.GroundTruthGenerator')
-    @patch('evidenceforge.generation.engine.ActivityGenerator')
-    @patch('evidenceforge.generation.engine.ZeekReporterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPacketFilterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPeEmitter')
-    @patch('evidenceforge.generation.engine.ZeekOcspEmitter')
-    @patch('evidenceforge.generation.engine.ZeekX509Emitter')
-    @patch('evidenceforge.generation.engine.ZeekWeirdEmitter')
-    @patch('evidenceforge.generation.engine.ZeekNtpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDhcpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekFilesEmitter')
-    @patch('evidenceforge.generation.engine.ZeekSslEmitter')
-    @patch('evidenceforge.generation.engine.ZeekHttpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDnsEmitter')
-    @patch('evidenceforge.generation.engine.ZeekEmitter')
-    @patch('evidenceforge.generation.engine.WindowsEventEmitter')
-    @patch('evidenceforge.generation.engine.SysmonEventEmitter')
-    @patch('evidenceforge.generation.engine.load_format')
+    @patch('evidenceforge.generation.engine.core.GroundTruthGenerator')
+    @patch('evidenceforge.generation.engine.core.ActivityGenerator')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekReporterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPacketFilterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPeEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekOcspEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekX509Emitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekWeirdEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekNtpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDhcpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekFilesEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekSslEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekHttpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDnsEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.WindowsEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.SysmonEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.load_format')
     def test_generate_skips_ground_truth_without_malicious_events(
         self, mock_load_format, mock_sysmon, mock_windows, mock_zeek, mock_zeek_dns, mock_zeek_http, mock_zeek_ssl, mock_zeek_files, mock_zeek_dhcp, mock_zeek_ntp, mock_zeek_weird, mock_zeek_x509, mock_zeek_ocsp, mock_zeek_pe, mock_zeek_pf, mock_zeek_reporter,
         mock_activity_gen, mock_gt_gen, minimal_scenario, tmp_path
@@ -638,23 +638,23 @@ class TestGenerationEngine:
         # Ground truth generator should NOT be called
         assert not mock_gt_gen.called
 
-    @patch('evidenceforge.generation.engine.ActivityGenerator')
-    @patch('evidenceforge.generation.engine.ZeekReporterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPacketFilterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPeEmitter')
-    @patch('evidenceforge.generation.engine.ZeekOcspEmitter')
-    @patch('evidenceforge.generation.engine.ZeekX509Emitter')
-    @patch('evidenceforge.generation.engine.ZeekWeirdEmitter')
-    @patch('evidenceforge.generation.engine.ZeekNtpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDhcpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekFilesEmitter')
-    @patch('evidenceforge.generation.engine.ZeekSslEmitter')
-    @patch('evidenceforge.generation.engine.ZeekHttpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDnsEmitter')
-    @patch('evidenceforge.generation.engine.ZeekEmitter')
-    @patch('evidenceforge.generation.engine.WindowsEventEmitter')
-    @patch('evidenceforge.generation.engine.SysmonEventEmitter')
-    @patch('evidenceforge.generation.engine.load_format')
+    @patch('evidenceforge.generation.engine.core.ActivityGenerator')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekReporterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPacketFilterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPeEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekOcspEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekX509Emitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekWeirdEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekNtpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDhcpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekFilesEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekSslEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekHttpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDnsEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.WindowsEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.SysmonEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.load_format')
     def test_finalize_closes_emitters(
         self, mock_load_format, mock_sysmon, mock_windows, mock_zeek, mock_zeek_dns, mock_zeek_http, mock_zeek_ssl, mock_zeek_files, mock_zeek_dhcp, mock_zeek_ntp, mock_zeek_weird, mock_zeek_x509, mock_zeek_ocsp, mock_zeek_pe, mock_zeek_pf, mock_zeek_reporter, mock_activity_gen,
         minimal_scenario, tmp_path
@@ -677,23 +677,23 @@ class TestGenerationEngine:
         assert mock_windows_instance.close.called
         assert mock_zeek_instance.close.called
 
-    @patch('evidenceforge.generation.engine.ActivityGenerator')
-    @patch('evidenceforge.generation.engine.ZeekReporterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPacketFilterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPeEmitter')
-    @patch('evidenceforge.generation.engine.ZeekOcspEmitter')
-    @patch('evidenceforge.generation.engine.ZeekX509Emitter')
-    @patch('evidenceforge.generation.engine.ZeekWeirdEmitter')
-    @patch('evidenceforge.generation.engine.ZeekNtpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDhcpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekFilesEmitter')
-    @patch('evidenceforge.generation.engine.ZeekSslEmitter')
-    @patch('evidenceforge.generation.engine.ZeekHttpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDnsEmitter')
-    @patch('evidenceforge.generation.engine.ZeekEmitter')
-    @patch('evidenceforge.generation.engine.WindowsEventEmitter')
-    @patch('evidenceforge.generation.engine.SysmonEventEmitter')
-    @patch('evidenceforge.generation.engine.load_format')
+    @patch('evidenceforge.generation.engine.core.ActivityGenerator')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekReporterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPacketFilterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPeEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekOcspEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekX509Emitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekWeirdEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekNtpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDhcpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekFilesEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekSslEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekHttpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDnsEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.WindowsEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.SysmonEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.load_format')
     def test_progress_callback_invoked(
         self, mock_load_format, mock_sysmon, mock_windows, mock_zeek, mock_zeek_dns, mock_zeek_http, mock_zeek_ssl, mock_zeek_files, mock_zeek_dhcp, mock_zeek_ntp, mock_zeek_weird, mock_zeek_x509, mock_zeek_ocsp, mock_zeek_pe, mock_zeek_pf, mock_zeek_reporter, mock_activity_gen,
         minimal_scenario, tmp_path
@@ -723,23 +723,23 @@ class TestGenerationEngine:
         assert len(phase_starts) > 0
         assert len(phase_ends) > 0
 
-    @patch('evidenceforge.generation.engine.ActivityGenerator')
-    @patch('evidenceforge.generation.engine.ZeekReporterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPacketFilterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPeEmitter')
-    @patch('evidenceforge.generation.engine.ZeekOcspEmitter')
-    @patch('evidenceforge.generation.engine.ZeekX509Emitter')
-    @patch('evidenceforge.generation.engine.ZeekWeirdEmitter')
-    @patch('evidenceforge.generation.engine.ZeekNtpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDhcpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekFilesEmitter')
-    @patch('evidenceforge.generation.engine.ZeekSslEmitter')
-    @patch('evidenceforge.generation.engine.ZeekHttpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDnsEmitter')
-    @patch('evidenceforge.generation.engine.ZeekEmitter')
-    @patch('evidenceforge.generation.engine.WindowsEventEmitter')
-    @patch('evidenceforge.generation.engine.SysmonEventEmitter')
-    @patch('evidenceforge.generation.engine.load_format')
+    @patch('evidenceforge.generation.engine.core.ActivityGenerator')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekReporterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPacketFilterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPeEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekOcspEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekX509Emitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekWeirdEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekNtpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDhcpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekFilesEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekSslEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekHttpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDnsEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.WindowsEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.SysmonEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.load_format')
     def test_progress_callback_not_required(
         self, mock_load_format, mock_sysmon, mock_windows, mock_zeek, mock_zeek_dns, mock_zeek_http, mock_zeek_ssl, mock_zeek_files, mock_zeek_dhcp, mock_zeek_ntp, mock_zeek_weird, mock_zeek_x509, mock_zeek_ocsp, mock_zeek_pe, mock_zeek_pf, mock_zeek_reporter, mock_activity_gen,
         minimal_scenario, tmp_path
@@ -770,23 +770,23 @@ class TestGenerationEngine:
         assert id2 == id1 + 1
         assert id3 == id2 + 1
 
-    @patch('evidenceforge.generation.engine.ActivityGenerator')
-    @patch('evidenceforge.generation.engine.ZeekReporterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPacketFilterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPeEmitter')
-    @patch('evidenceforge.generation.engine.ZeekOcspEmitter')
-    @patch('evidenceforge.generation.engine.ZeekX509Emitter')
-    @patch('evidenceforge.generation.engine.ZeekWeirdEmitter')
-    @patch('evidenceforge.generation.engine.ZeekNtpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDhcpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekFilesEmitter')
-    @patch('evidenceforge.generation.engine.ZeekSslEmitter')
-    @patch('evidenceforge.generation.engine.ZeekHttpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDnsEmitter')
-    @patch('evidenceforge.generation.engine.ZeekEmitter')
-    @patch('evidenceforge.generation.engine.WindowsEventEmitter')
-    @patch('evidenceforge.generation.engine.SysmonEventEmitter')
-    @patch('evidenceforge.generation.engine.load_format')
+    @patch('evidenceforge.generation.engine.core.ActivityGenerator')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekReporterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPacketFilterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPeEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekOcspEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekX509Emitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekWeirdEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekNtpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDhcpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekFilesEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekSslEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekHttpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDnsEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.WindowsEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.SysmonEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.load_format')
     def test_execute_storyline_event_logon_type(
         self, mock_load_format, mock_sysmon, mock_windows, mock_zeek, mock_zeek_dns, mock_zeek_http, mock_zeek_ssl, mock_zeek_files, mock_zeek_dhcp, mock_zeek_ntp, mock_zeek_weird, mock_zeek_x509, mock_zeek_ocsp, mock_zeek_pe, mock_zeek_pf, mock_zeek_reporter, mock_activity_gen,
         scenario_with_storyline, tmp_path
@@ -813,23 +813,23 @@ class TestGenerationEngine:
         call_args = mock_activity_instance.generate_logon.call_args
         assert call_args[1]['logon_type'] == 3
 
-    @patch('evidenceforge.generation.engine.ActivityGenerator')
-    @patch('evidenceforge.generation.engine.ZeekReporterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPacketFilterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPeEmitter')
-    @patch('evidenceforge.generation.engine.ZeekOcspEmitter')
-    @patch('evidenceforge.generation.engine.ZeekX509Emitter')
-    @patch('evidenceforge.generation.engine.ZeekWeirdEmitter')
-    @patch('evidenceforge.generation.engine.ZeekNtpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDhcpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekFilesEmitter')
-    @patch('evidenceforge.generation.engine.ZeekSslEmitter')
-    @patch('evidenceforge.generation.engine.ZeekHttpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDnsEmitter')
-    @patch('evidenceforge.generation.engine.ZeekEmitter')
-    @patch('evidenceforge.generation.engine.WindowsEventEmitter')
-    @patch('evidenceforge.generation.engine.SysmonEventEmitter')
-    @patch('evidenceforge.generation.engine.load_format')
+    @patch('evidenceforge.generation.engine.core.ActivityGenerator')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekReporterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPacketFilterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPeEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekOcspEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekX509Emitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekWeirdEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekNtpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDhcpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekFilesEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekSslEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekHttpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDnsEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.WindowsEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.SysmonEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.load_format')
     def test_execute_storyline_event_connection_validation(
         self, mock_load_format, mock_sysmon, mock_windows, mock_zeek, mock_zeek_dns, mock_zeek_http, mock_zeek_ssl, mock_zeek_files, mock_zeek_dhcp, mock_zeek_ntp, mock_zeek_weird, mock_zeek_x509, mock_zeek_ocsp, mock_zeek_pe, mock_zeek_pf, mock_zeek_reporter, mock_activity_gen,
         scenario_with_storyline, tmp_path
@@ -858,23 +858,23 @@ class TestGenerationEngine:
         call_args = mock_activity_instance.generate_connection.call_args
         assert call_args[1]['dst_ip'] == "159.65.43.201"
 
-    @patch('evidenceforge.generation.engine.ActivityGenerator')
-    @patch('evidenceforge.generation.engine.ZeekReporterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPacketFilterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPeEmitter')
-    @patch('evidenceforge.generation.engine.ZeekOcspEmitter')
-    @patch('evidenceforge.generation.engine.ZeekX509Emitter')
-    @patch('evidenceforge.generation.engine.ZeekWeirdEmitter')
-    @patch('evidenceforge.generation.engine.ZeekNtpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDhcpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekFilesEmitter')
-    @patch('evidenceforge.generation.engine.ZeekSslEmitter')
-    @patch('evidenceforge.generation.engine.ZeekHttpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDnsEmitter')
-    @patch('evidenceforge.generation.engine.ZeekEmitter')
-    @patch('evidenceforge.generation.engine.WindowsEventEmitter')
-    @patch('evidenceforge.generation.engine.SysmonEventEmitter')
-    @patch('evidenceforge.generation.engine.load_format')
+    @patch('evidenceforge.generation.engine.core.ActivityGenerator')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekReporterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPacketFilterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPeEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekOcspEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekX509Emitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekWeirdEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekNtpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDhcpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekFilesEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekSslEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekHttpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDnsEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.WindowsEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.SysmonEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.load_format')
     def test_generate_user_activity_uses_primary_system(
         self, mock_load_format, mock_sysmon, mock_windows, mock_zeek, mock_zeek_dns, mock_zeek_http, mock_zeek_ssl, mock_zeek_files, mock_zeek_dhcp, mock_zeek_ntp, mock_zeek_weird, mock_zeek_x509, mock_zeek_ocsp, mock_zeek_pe, mock_zeek_pf, mock_zeek_reporter, mock_activity_gen,
         minimal_scenario, tmp_path
@@ -907,23 +907,23 @@ class TestGenerationEngine:
         call_args = mock_activity_instance.execute_baseline_activity.call_args
         assert call_args[1]['system'].hostname == "TEST-01"
 
-    @patch('evidenceforge.generation.engine.ActivityGenerator')
-    @patch('evidenceforge.generation.engine.ZeekReporterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPacketFilterEmitter')
-    @patch('evidenceforge.generation.engine.ZeekPeEmitter')
-    @patch('evidenceforge.generation.engine.ZeekOcspEmitter')
-    @patch('evidenceforge.generation.engine.ZeekX509Emitter')
-    @patch('evidenceforge.generation.engine.ZeekWeirdEmitter')
-    @patch('evidenceforge.generation.engine.ZeekNtpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDhcpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekFilesEmitter')
-    @patch('evidenceforge.generation.engine.ZeekSslEmitter')
-    @patch('evidenceforge.generation.engine.ZeekHttpEmitter')
-    @patch('evidenceforge.generation.engine.ZeekDnsEmitter')
-    @patch('evidenceforge.generation.engine.ZeekEmitter')
-    @patch('evidenceforge.generation.engine.WindowsEventEmitter')
-    @patch('evidenceforge.generation.engine.SysmonEventEmitter')
-    @patch('evidenceforge.generation.engine.load_format')
+    @patch('evidenceforge.generation.engine.core.ActivityGenerator')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekReporterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPacketFilterEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekPeEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekOcspEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekX509Emitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekWeirdEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekNtpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDhcpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekFilesEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekSslEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekHttpEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekDnsEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.ZeekEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.WindowsEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.SysmonEventEmitter')
+    @patch('evidenceforge.generation.engine.emitter_setup.load_format')
     def test_execute_storyline_skips_missing_actor(
         self, mock_load_format, mock_sysmon, mock_windows, mock_zeek, mock_zeek_dns, mock_zeek_http, mock_zeek_ssl, mock_zeek_files, mock_zeek_dhcp, mock_zeek_ntp, mock_zeek_weird, mock_zeek_x509, mock_zeek_ocsp, mock_zeek_pe, mock_zeek_pf, mock_zeek_reporter, mock_activity_gen,
         scenario_with_storyline, tmp_path
